pi/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.1
ollama==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
sounddevice==0.5.3
srt==3.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vosk==0.3.45
websockets==16.0
